--- a/staging/src/k8s.io/pod-security-admission/metrics/metrics.go
+++ b/staging/src/k8s.io/pod-security-admission/metrics/metrics.go
@@ -17,7 +17,15 @@ limitations under the License.
 package metrics
 
 import (
+	"strconv"
+	"strings"
+	"sync"
+
+	admissionv1 "k8s.io/api/admission/v1"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/component-base/metrics"
+	"k8s.io/component-base/metrics/legacyregistry"
 	"k8s.io/pod-security-admission/api"
 )
 
@@ -29,45 +37,67 @@ const (
 	DecisionDeny  = "deny"  // Policy evaluated, request denied
 )
 
-var (
-	SecurityEvaluation = metrics.NewCounterVec(
-		&metrics.CounterOpts{
-			Name:           "pod_security_evaluations_total",
-			Help:           "Counter of pod security evaluations.",
-			StabilityLevel: metrics.ALPHA,
-		},
-		[]string{"decision", "policy_level", "policy_version", "mode", "operation", "resource", "subresource"},
-	)
-
-	Registry = metrics.NewKubeRegistry()
-)
-
 type Decision string
 type Mode string
 
 type Recorder interface {
-	RecordEvaluation(decision Decision, policy api.LevelVersion, evalMode Mode, attrs api.Attributes)
+	RecordEvaluation(Decision, api.LevelVersion, Mode, api.Attributes)
+}
+
+var defaultRecorder = NewPrometheusRecorder(api.GetAPIVersion())
+
+func DefaultRecorder() Recorder {
+	return defaultRecorder
+}
+
+// MustRegister registers the global DefaultMetrics against the legacy registry.
+func LegacyMustRegister() {
+	defaultRecorder.MustRegister(legacyregistry.MustRegister)
 }
 
 type PrometheusRecorder struct {
 	apiVersion api.Version
+
+	evaluationsCounter *metrics.CounterVec
+
+	registerOnce sync.Once
 }
 
-func init() {
-	Registry.MustRegister(SecurityEvaluation)
-}
+var _ Recorder = &PrometheusRecorder{}
 
 func NewPrometheusRecorder(version api.Version) *PrometheusRecorder {
-	return &PrometheusRecorder{apiVersion: version}
+	evaluationsCounter := metrics.NewCounterVec(
+		&metrics.CounterOpts{
+			Name:           "pod_security_evaluations_total",
+			Help:           "Number of policy evaluations that occurred, not counting ignored or exempt requests.",
+			StabilityLevel: metrics.ALPHA,
+		},
+		[]string{"decision", "policy_level", "policy_version", "mode", "request_operation", "resource", "subresource"},
+	)
+
+	return &PrometheusRecorder{
+		apiVersion:         version,
+		evaluationsCounter: evaluationsCounter,
+	}
 }
 
-func (r PrometheusRecorder) RecordEvaluation(decision Decision, policy api.LevelVersion, evalMode Mode, attrs api.Attributes) {
+func (r *PrometheusRecorder) MustRegister(registerFunc func(...metrics.Registerable)) {
+	r.registerOnce.Do(func() {
+		registerFunc(r.evaluationsCounter)
+	})
+}
+
+func (r *PrometheusRecorder) Reset() {
+	r.evaluationsCounter.Reset()
+}
+
+func (r *PrometheusRecorder) RecordEvaluation(decision Decision, policy api.LevelVersion, evalMode Mode, attrs api.Attributes) {
 	dec := string(decision)
-	operation := string(attrs.GetOperation())
-	resource := attrs.GetResource().String()
+	operation := operationLabel(attrs.GetOperation())
+	resource := resourceLabel(attrs.GetResource())
 	subresource := attrs.GetSubresource()
+
 	var version string
-	if policy.Valid() {
 		if policy.Version.Latest() {
 			version = "latest"
 		} else {
@@ -77,7 +107,31 @@ func (r PrometheusRecorder) RecordEvaluation(decision Decision, policy api.Level
 				version = "future"
 			}
 		}
-		SecurityEvaluation.WithLabelValues(dec, string(policy.Level),
+
+		r.evaluationsCounter.WithLabelValues(dec, string(policy.Level),
 			version, string(evalMode), operation, resource, subresource).Inc()
+}
+
+func resourceLabel(resource schema.GroupVersionResource) string {
+	switch resource.GroupResource() {
+	case corev1.Resource("pods"):
+		return "pod"
+	case corev1.Resource("namespace"):
+		return "namespace"
+	default:
+		// Assume any other resource is a valid input to pod-security, and therefore a controller.
+		return "controller"
+	}
+}
+
+func operationLabel(op admissionv1.Operation) string {
+	switch op {
+	case admissionv1.Create:
+		return "create"
+	case admissionv1.Update:
+		return "update"
+	default:
+		// This is a slower operation, but never used in the default implementation.
+		return strings.ToLower(string(op))
 	}
 }