@tallclair 2021-11-01
Rebased. Also had to update the test to check specifically for `audit-violations` in the annotations to construct the metrics expectations.